Move Mock to a dedicated package and rename interface
To avoid inclusion of Mock instance outside of test in users’ packages.
Dependencies introduced by mock will not be compiled for users’ runtime
